refactor: 패키지 위치 변경
#416 피드백 반영-1

- JPA Repository 모두 repository 패키지로 이동
- 서비스 클래스가 잘못된 위치에 있어 변경함
