+++ src/main/java/nextstep/subway/repository/LineRepository.java
@@ -1,5 +1,6 @@
-package nextstep.subway.domain;
+package nextstep.subway.repository;
 
+import nextstep.subway.domain.Line;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
